pyproject.toml
[project]
name = "kimi-cli"
version = "0.59"
description = "Kimi CLI is your next CLI agent."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agent-client-protocol==0.6.3",
    "aiofiles==25.1.0",
    "aiohttp==3.13.2",
    "typer==0.20.0",
    "kosong==0.28.0",
    "loguru==0.7.3",
    "prompt-toolkit==3.0.52",
    "pillow==12.0.0",
    "pyyaml==6.0.3",
    "rich==14.2.0",
    "ripgrepy==2.2.0",
    "streamingjson==0.0.5",
    "trafilatura==2.0.0",
    "tenacity==9.1.2",
    "fastmcp==2.12.5",
    "pydantic==2.12.5",
    "httpx[socks]==0.28.1",
]

[dependency-groups]
dev = [
    "inline-snapshot[black]>=0.31.1",
    "pyinstaller>=6.16.0",
    "pyright>=1.1.407",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.14.6",
]

[build-system]
requires = ["uv_build>=0.8.5,<0.9.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "kimi_cli"
source-exclude = ["examples/**/*", "tests/**/*", "src/kimi_cli/deps/**/*"]

[project.scripts]
kimi = "kimi_cli.cli:cli"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # Pyflakes
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "I",   # isort
]

[tool.typos.files]
extend-exclude = ["kimi.spec", "pyinstaller.py"]

[tool.typos.default.extend-words]
datas = "datas"
